Rename database connection pools view
